+++ src/components/layout/SideTower/SideTower.jsx
@@ -9,6 +9,12 @@ import {
 } from './SideTower.styles';
 import { ControlTower, WorkbenchDrawer, SideTowerNotes } from '../../tower';
 
+/**
+ * Tour latérale contenant les contrôles et outils
+ * @renders ControlTower
+ * @renders WorkbenchDrawer
+ * @renders SideTowerNotes
+ */
 const SideTower = () => {
   return (
     <TowerContainer>